+++ examples/exampleinplayservice.py
@@ -1,4 +1,3 @@
-import os
 import logging
 
 import betfairlightweight
@@ -13,8 +12,9 @@ timeline will give update details.
 logging.basicConfig(level=logging.INFO)  # change to DEBUG to see log all updates
 
 # create trading instance
-username = os.environ.get("username")
-trading = betfairlightweight.APIClient(username)
+trading = betfairlightweight.APIClient("username", "password", app_key="appKey")
+
+# login
 trading.login()
 
 # update