+++ HISTORY.rst
@@ -19,6 +19,7 @@ Release History
 - __version__ added to user agent
 - raw requests Response added to objects
 - elapsed_time now uses time() rather than datetime
+- session can be passed to client
 
 **Bug Fixes**